requirements.txt
cassandra-driver==3.24.0
pyqrcode==1.2.1
onetimepass==1.0.1
flask-wtf==0.14.3
flask-bootstrap==3.3.7.1
flask-sqlalchemy==2.4.4
flask-paginate==0.7.0
flask-login==0.5.0